Simplify the formula registration process
Currently, the `Fontist::Formulas` modules was taking care of
formula registration but since it's a module then we had some
limitation, and also not so direct interface.

This commit moves these to the `Fontist::Formula` interface,
now any instantiation of this class will ensure the formula
registration. This also use some ruby specific feature to
cache the registering behavior, so it should be bit of an
optimization.
